Initial Upload of the A6K and Nimble PSTK version 3.4.1
This initial upload contains the Nimble PowerShell Toolkit 3.4.1 which is the basis for the Nimble Storage and Alletra 6000 PowerShell commands; The only files not included are the license as the license for the new toolkit will be the BSD Clause 3 license. Additional, to load THIS powershell toolkit, run the following command from a powershell prompt in this directory : Import-Module .\HPEAlletra6000andNimbleStoragePowerShellToolkit.psd1. 
When the 3PAR/Primera/Alletra code is added to the same folder, there are no files that will conflict, and it will require a different Import-Module command to run.