Adding the 3PAR/Primera/A9K PSTK v3.1 as a basis
This is the initial upload of the HPE 3PAR / Primera / Alletra 9000 PowerShell Toolkit version 3.1.0 which is the basis for a new combined Toolkit. 
To use this module from the main directory, run the following command : Import-Module .\HPEStoragePowerShellToolkit 

This toolkit is installed into the same directory as the HPE Nimble Storage Toolkit since neither had files that conflict with the others.